Extract Configuration from XcodeList
